added mui Stepper to indicate progress
