+++ src/components/Confirmation/Confirmation.jsx
@@ -1,13 +1,15 @@
 import { useDispatch } from "react-redux";
 import { useHistory } from "react-router-dom";
 
-export default function Confirmation() {
+export default function Confirmation({ funcsFromStepper }) {
   const dispatch = useDispatch();
 
   const history = useHistory();
 
   const handleClick = () => {
     dispatch({ type: "CLEAR_FEEDBACK" });
+    // reset the Stepper
+    funcsFromStepper.handleReset();
     // TODO route the user back to feeling page
     history.push("/");
   };
